requirements.txt
# Core Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Financial Data
yfinance>=0.1.70

# Web Scraping
requests>=2.26.0
beautifulsoup4>=4.10.0
lxml>=4.9.0

# Interactive Brokers API
ib_insync>=0.9.70

# Database
sqlalchemy>=1.4.0

# Configuration
python-dotenv>=0.19.0

# Scheduling
schedule>=1.1.0
APScheduler>=3.9.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Logging & Monitoring
python-json-logger>=2.0.0

# Data Visualization (for notebooks)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0

# Jupyter for analysis
jupyter>=1.0.0
notebook>=6.4.0

# Additional utilities
tqdm>=4.62.0  # Progress bars
pytz>=2021.3  # Timezone handling
